Split copiers into separate files
