+++ lib/serialtilescopy.js
@@ -1,11 +1,10 @@
-var url = require('url');
-var zlib = require('zlib');
 var fs = require('fs');
 var util = require('util');
-var S3 = require('tilelive-s3');
 var tilelive = require('tilelive');
+var zlib = require('zlib');
+var url = require('url');
 var progress = require('progress-stream');
-var queue = require('queue-async');
+var S3 = require('tilelive-s3');
 var ValidationStream = require('mapbox-upload-validate/lib/validators/serialtiles').ValidationStream;
 
 function serialtiles(srcUri, s3urlTemplate, options, callback) {
@@ -82,53 +81,4 @@ function serialtiles(srcUri, s3urlTemplate, options, callback) {
   }
 }
 
-function tilelivecopy(srcUri, s3url, options, callback) {
-  var scheme = 'scanline';
-  var protocol = url.parse(srcUri).protocol;
-  if (protocol === 'mbtiles:') scheme = 'list';
-  if (protocol === 'omnivore:') scheme = 'pyramid';
-  options.type = scheme;
-
-  queue()
-    .defer(tilelive.load, srcUri)
-    .defer(tilelive.load, s3url)
-    .await(function(err, src, dst) {
-      function close(err) {
-        var srcClose = src && typeof src.close === 'function' ?
-          src.close.bind(src) : function(cb) { cb(); };
-        var dstClose = dst && typeof dst.close === 'function' ?
-          dst.close.bind(dst) : function(cb) { cb(); };
-
-        queue()
-          .defer(srcClose)
-          .defer(dstClose)
-          .await(function(closeErr) {
-            callback(err || closeErr);
-          });
-      }
-
-      if (err) {
-        err.code = 'EINVALID';
-        return close(err);
-      }
-
-      src.getInfo(function(err, info) {
-        if (err) {
-          err.code = 'EINVALID';
-          return close(err);
-        }
-
-        options.minzoom = options.minzoom || info.minzoom;
-        options.maxzoom = options.maxzoom || info.maxzoom;
-        options.bounds = options.bounds || info.bounds;
-        if (scheme === 'list') options.listStream = src.createZXYStream();
-
-        tilelive.copy(src, dst, options, close);
-      });
-    });
-}
-
-module.exports = {
-  tilelive: tilelivecopy,
-  serialtiles: serialtiles
-};
+module.exports = serialtiles;
