+++ test/copy.tilelive.test.js
@@ -1,7 +1,7 @@
 var test = require('tape');
 var path = require('path');
 var crypto = require('crypto');
-var tileliveCopy = require('../lib/copy').tilelive;
+var tileliveCopy = require('../lib/tilelivecopy');
 var tilelive = require('tilelive');
 var AWS = require('aws-sdk');
 var s3urls = require('s3urls');
